refactor: Product ID 타입을 String에서 Int로 변경 및 데이터베이스 버전 업그레이드

+++ app/src/main/java/woowacourse/shopping/data/mock/MockProductResponseFactory.kt
@@ -20,7 +20,7 @@ object MockProductResponseFactory {
 
     private fun Product.toResponseDto(): ProductResponseDto =
         ProductResponseDto(
-            id = id.toInt(),
+            id = id,
             imageUrl = imageUrl.value,
             name = name.value,
             price = price.value,
